SahYatri-DetectionAI/requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
ultralytics==8.0.196
python-multipart==0.0.6
requests==2.31.0
pillow==10.1.0
python-dotenv==1.0.0
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
numpy==1.24.3  